timetable_miner/requirements.txt
tabula-py
Flask